ndb - fix compiler warnings  - remove unimplemented ArrayPool::load causing vc++ to complain when instantiating  - Use Uint64 variable for incrementin GlobalData.JobLap  - Remove unecessary forward declaration of IPCConfig

+++ storage/ndb/src/kernel/vm/FastScheduler.cpp
@@ -105,7 +105,7 @@ FastScheduler::doJob()
       globalData.incrementWatchDogCounter(1);
       if (reg_bnr > 0) {
         Uint32 tJobCounter = globalData.JobCounter;
-        Uint32 tJobLap = globalData.JobLap;
+        Uint64 tJobLap = globalData.JobLap;
         SimulatedBlock* b = globalData.getBlock(reg_bnr);
         theJobPriority[tJobCounter] = (Uint8)tHighPrio;
         globalData.JobCounter = (tJobCounter + 1) & 4095;
